+++ podwodny_swiat_tabliczki_mnozenia/Assets/Scenes/NameScene.unity
@@ -134,7 +134,7 @@ GameObject:
   - component: {fileID: 253496682}
   - component: {fileID: 253496681}
   m_Layer: 5
-  m_Name: Button
+  m_Name: Button_Dalej
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -441,7 +441,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 3ce14fbd9ca5757409df68e2c362980d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  player_name: 
   inputField: {fileID: 1060006346}
 --- !u!1 &1060006346
 GameObject:
@@ -758,7 +757,7 @@ GameObject:
   - component: {fileID: 1624182894}
   - component: {fileID: 1624182893}
   m_Layer: 5
-  m_Name: Text
+  m_Name: Text_PodajImie
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -836,7 +835,7 @@ GameObject:
   - component: {fileID: 1714647980}
   - component: {fileID: 1714647979}
   m_Layer: 5
-  m_Name: InputField
+  m_Name: Input_Imie
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
