+++ podwodny_swiat_tabliczki_mnozenia/Assets/Scripts/FishController.cs
@@ -1,0 +1,29 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class FishController : MonoBehaviour
+{
+    // Start is called before the first frame update
+    public float moveSpeed = 1.0f;
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        GetComponent<Rigidbody2D>().velocity = new Vector2(-moveSpeed, GetComponent<Rigidbody2D>().velocity.y);
+
+        if (Input.GetMouseButtonDown(0))
+        {
+            RaycastHit2D rayHit = Physics2D.GetRayIntersection(Camera.main.ScreenPointToRay(Input.mousePosition));
+            if (rayHit)
+            {
+                Debug.Log( rayHit.transform.name + " clicked");
+            }
+            
+        }
+    }
+}
