+++ podwodny_swiat_tabliczki_mnozenia/Assets/Scripts/FishController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 00a130a02e728004c94609ff8d1eb0d6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
