+++ podwodny_swiat_tabliczki_mnozenia/Assets/Scripts/GetPlayersName.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3ce14fbd9ca5757409df68e2c362980d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
