--- podwodny_swiat_tabliczki_mnozenia/Assets/Scripts/Name.cs
@@ -1,19 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.UI;
-
-public class Name : MonoBehaviour
-{
-
-    public string player_name;
-    public GameObject inputField; 
-
-
-    public void Button_Click()
-    {
-        player_name = inputField.GetComponent<Text>().text;
-        Debug.Log( "Witaj " + player_name + " w podwodnym świecie tabliczki mnożenia!");
-    }
-
-}
