--- podwodny_swiat_tabliczki_mnozenia/Assets/Scritps.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8fc4a94bba6b70d48aeb2c0f7f791aee
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
